split heuristic file
